Friend page initial implementation, bug fixes.
Added functionality to manage the list of friends added to the bill: includes setting their name and items that they are paying for.
Fixed a bug where the app would crash when attempting to exit using the back button. This involves changing the layout of the icons on the bar and removing the empty index.tsx from being loaded.
Fixed a bug where AsyncStorage wasn't storing the state properly onto the device.