Completed home screen and added share function.
Added the display elements and functions for the home screen to show subtotals, totals, and the breakdown of items and amount owed by each person.
Added getters to the Item class to make calculating and grabbing tax or price amounts easier.
Modified functions in each page to use the getters rather than the previous hardcoded calculations.
Updated the Item class to include a separate PST variable to separate between 7% PST and 10% PST (for things like alcohol).
Updated the state context to properly use state the way React Native intended it.